+++ admin/documentation/code-samples/custom-template-header.php
@@ -14,11 +14,12 @@ if ( ! defined( 'ABSPATH' ) ) {
 
 ?>
 
-<textarea cols="70" rows="7" wrap="off" style="width: 100%;" class="code">
+<textarea cols="70" rows="8" wrap="off" style="width: 100%;" class="code">
 &lt;?php
 
 /**
  * Page Part Name: My Custom Template
+ * Page Part Image: images/page-parts/my-custom-template.png
  */
 
 ?&gt;</textarea>